refactor: move auxiliary functions to lib
