+++ src/app/api/map/route.ts
@@ -1,6 +1,6 @@
 import type {NextRequest} from 'next/server'
 
-import dbConnect from "@/lib/dbConnect";
+import dbConnect from "@/lib/utils/dbConnect";
 import {NextResponse} from "next/server";
 import models from "@/models";
