+++ src/app/api/shop/route.ts
@@ -1,5 +1,5 @@
 import type {NextRequest} from 'next/server'
-import dbConnect from "@/lib/dbConnect";
+import dbConnect from "@/lib/utils/dbConnect";
 import {NextResponse} from "next/server";
 import Shop from "@/models/Shop";
