--- src/components/footer/Footer..tsx
@@ -1,9 +1,0 @@
-const Footer = () => {
-    return (
-        <footer>
-
-        </footer>
-    );
-};
-
-export default Footer;
