+++ src/lib/constants/graphs/graph3.ts
