+++ src/lib/utils/dijkstra.ts
