+++ src/lib/utils/escapeHTML.ts
@@ -1,0 +1,10 @@
+const sanitizeHtmlString = (unsafe: string) => {
+    return unsafe
+        .replace(/&/g, '&amp;')
+        .replace(/</g, '&lt;')
+        .replace(/>/g, '&gt;')
+        .replace(/"/g, '&quot;')
+        .replace(/'/g, '&#039;');
+};
+
+export default sanitizeHtmlString
